+++ src/ons_pd_may_19.py
@@ -45,7 +45,7 @@ class ONSPostcodeDirectoryMay19(ONSPostcodeDirectory):
         ONSPostcodeDirectory.__init__(self, ons_pd_csv_path, load_data, ONSPostcodeDirectoryMay19.index_column, ONSPostcodeDirectoryMay19.fields, ONSPostcodeDirectoryMay19.data_types)
 
         # Folder within the ONS Postcode Directory archive holding names and codes files
-        self.NAMES_AND_CODES_FILE_LOCATION = self.settings["ONS Names and codes folder"]
+        NAMES_AND_CODES_FOLDER = self.settings["ONS Names and codes folder"]
 
         # Paths to all shapefiles within the Boundaries folder
         shapefile_paths = {
@@ -53,7 +53,7 @@ class ONSPostcodeDirectoryMay19(ONSPostcodeDirectory):
             "County": self.settings["Boundaries folder"] + r"Counties_and_Unitary_Authorities_December_2017_Generalised_Clipped_Boundaries_in_UK\Counties_and_Unitary_Authorities_December_2017_Generalised_Clipped_Boundaries_in_UK.shp",
             "Ward": self.settings["Boundaries folder"] + r"Wards_May_2019_Boundaries_UK_BSC\Wards_May_2019_Boundaries_UK_BSC.shp",
             "PCon": self.settings["Boundaries folder"] + r"Westminster_PCON_Dec_2017_Generalised_Clipped_UK\Westminster_Parliamentary_Constituencies_December_2017_Generalised_Clipped_Boundaries_in_the_UK.shp",
-            "LSOA": self.settings["Boundaries folder"] + r"Lower_Layer_Super_Output_Areas_December_2011_Generalised_Clipped__Boundaries_in_England_and_Wales\Lower_Layer_Super_Output_Areas_December_2011_Generalised_Clipped__Boundaries_in_England_and_Wales.shp",
+            "LSOA": self.settings["Boundaries folder"] + r"Lower Layer Super Output Areas (December 2011) Boundaries EW BGC\Lower Layer Super Output Areas (December 2011) Boundaries EW BGC.shp",
             "MSOA": self.settings["Boundaries folder"] + r"Middle_Layer_Super_Output_Areas_December_2011_Full_Clipped_Boundaries_in_England_and_Wales\Middle_Layer_Super_Output_Areas_December_2011_Full_Clipped_Boundaries_in_England_and_Wales.shp",
             "IZ": self.settings["Boundaries folder"] + r"SG_IntermediateZoneBdry_2011\SG_IntermediateZone_Bdry_2011.shp",
         }
@@ -62,42 +62,42 @@ class ONSPostcodeDirectoryMay19(ONSPostcodeDirectory):
             "lad": {
                 # Local Authority Districts
                 "name": "oslaua",
-                "codes": {"path": "LA_UA names and codes UK as at 12_19.csv", "key": "LAD19CD", "key_type": "object", "name": "LAD19NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "LA_UA names and codes UK as at 12_19.csv", "key": "LAD19CD", "key_type": "object", "name": "LAD19NM"},
                 "boundary": {"shapefile": shapefile_paths["LADs"], "key": 'lad19cd', "name": 'lad19nm', },
                 "age_profile": {"path": "lad_by_age.csv", "key": "Code"},
             },
             "cty": {
                 # Counties
                 "name": "oslaua",
-                "codes": {"path": "LA_UA names and codes UK as at 12_19.csv", "key": "LAD19CD", "key_type": "object", "name": "LAD19NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "LA_UA names and codes UK as at 12_19.csv", "key": "LAD19CD", "key_type": "object", "name": "LAD19NM"},
                 "boundary": {"shapefile": shapefile_paths["County"], "key": 'ctyua17cd', "name": 'ctyua17nm', },
                 "age_profile": {"path": None, "key": None},
             },
             "osward": {
                 # Council Wards
                 "name": "osward",
-                "codes": {"path": "Ward names and codes UK as at 05_19_NSPD.csv", "key": "WD19CD", "key_type": "object", "name": "WD19NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "Ward names and codes UK as at 05_19_NSPD.csv", "key": "WD19CD", "key_type": "object", "name": "WD19NM"},
                 "boundary": {"shapefile": shapefile_paths["Ward"], "key": 'wd19cd', "name": 'wd19nm', },
                 "age_profile": {"path": None, "key": None},
             },
             "pcon": {
                 # Parliamentary Constituencies
                 "name": "pcon",
-                "codes": {"path": "Westminster Parliamentary Constituency names and codes UK as at 12_14.csv", "key": "PCON14CD", "key_type": "object", "name": "PCON14NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "Westminster Parliamentary Constituency names and codes UK as at 12_14.csv", "key": "PCON14CD", "key_type": "object", "name": "PCON14NM"},
                 "boundary": {"shapefile": shapefile_paths["PCon"], "key": 'pcon17cd', "name": "pcon17nm", },
                 "age_profile": {"path": "pcon_by_age.csv", "key": "PCON11CD"},
             },
             "lsoa": {
                 # Lower Level Super Output Areas
                 "name": "lsoa11",
-                "codes": {"path": "LSOA (2011) names and codes UK as at 12_12.csv", "key": "LSOA11CD", "key_type": "object", "name": "LSOA11NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "LSOA (2011) names and codes UK as at 12_12.csv", "key": "LSOA11CD", "key_type": "object", "name": "LSOA11NM"},
                 "boundary": {"shapefile": shapefile_paths["LSOA"], "key": 'lsoa11cd', "name": 'lsoa11nm', },
                 "age_profile": {"path": None, "key": None},
             },
             "msoa": {
                 # Middle Layer Super Output Areas
                 "name": "msoa",
-                "codes": {"path": "MSOA (2011) names and codes UK as at 12_12.csv", "key": "MSOA11CD", "key_type": "object", "name": "MSOA11NM"},
+                "codes": {"path": NAMES_AND_CODES_FOLDER + "MSOA (2011) names and codes UK as at 12_12.csv", "key": "MSOA11CD", "key_type": "object", "name": "MSOA11NM"},
                 "boundary": {"shapefile": shapefile_paths["MSOA"], "key": 'msoa11cd', "name": None, },
                 "age_profile": {"path": None, "key": None},
             },
